Update crc32 call to use seed value and negate result
Originally CryptoSwift's CRC32 used a seed of all 1's (0xFF...), and
negated the final result (xor with 0xFF... again), but this wasn't the
CRC32 used by the tokens.  Originally I went into the CryptoSwitch
submodule and tweaked the CRC32, but this caused problems when other
people tried to use the repo and the submodule had a sha value that
wasn't in the remote repo.  With the upgrade to CryptoSwitch 0.3.1, the
CRC32 function now supports a seed parameter, so my tweaks aren't needed
and the correct CRC32 for the tokens can still be calculated.